Adicionando operações do basicas do Fornecedor com espaço para ajustes após a implementação do fornecimento e medicamentos.

+++ src/main/java/com/farmaciapaguemais/demo/repositories/FornecedorRepository.java
@@ -3,5 +3,5 @@ package com.farmaciapaguemais.demo.repositories;
 import com.farmaciapaguemais.demo.entities.Fornecedor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FornecedorRepository extends JpaRepository<Fornecedor, Integer> {
+public interface FornecedorRepository extends JpaRepository<Fornecedor, Long> {
 }
